made the previous commit compile. see below for more details:
Using the -disable-class-remapping option will allow people to see
some test cases passing if they are having trouble with the remapper.
However the remapper is required for some testcases that use the
synchronized keyword in Java or the Random class. With this new
option enabled those test cases will deadlock on the GPU.